+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - This CHANGELOG.md file.
 - `\Ely\Mojang\Api::setClient()` method to override default HTTP client.
+- [API Status](https://wiki.vg/Mojang_API#API_Status) endpoint.
 
 ### Changed
 - The constructor no longer has arguments.